// src/app/service/api/sub-title/interfaces/req-sub-title.interface.ts
export interface ReqCreateSubTitle {
        titleId: number
        subject: string
}

// export interface ResCreateSubTitle {
//         resCode: string
//         msg: string
//         resData: ResCreateSubTitleData
// }

// export interface ResCreateSubTitleData {
//         id: number
//         subject: string
//         titleId: number
// }